requirements/base.txt
Django>=1.9
django-environ>=0.4.0
django-braces>=1.8.1
django-crispy-forms>=1.5.2
django-admin-bootstrapped>=2.5.6
easy-thumbnails>=2.3
django-authtools
django-haystack==2.4.1
Whoosh==2.7.0
